Refactor SiwG to use Web_Tag class.
